fixed display bug, config ,alternative networking
display wouldnt actually update when using Shift+R and upon exiting config. It would print out current weather alerts instead due to a failure to update the alerts list when calling get_alerts()

made it so when editing a config variable it returns you back to the configuration screen instead of to the alerts page

in /servers there is a new file called "srvweather.py" which if you host either locally or on a server your device can talk to, it will automatically fetch and backup weather alerts making it far faster than using the normal method via weather.gov since you skip using https and downloading the full file. the server parses it for you.